+++ R/cassettes.R
@@ -21,11 +21,9 @@ insert_cassette <- function(
   dir = NULL,
   record = NULL,
   match_requests_on = NULL,
-  allow_playback_repeats = FALSE,
   serialize_with = NULL,
   preserve_exact_body_bytes = NULL,
   re_record_interval = NULL,
-  clean_outdated_http_interactions = NULL,
   warn_on_empty = NULL
 ) {
   if (vcr_turned_off()) {
@@ -41,11 +39,9 @@ insert_cassette <- function(
     dir = dir,
     record = record,
     match_requests_on = match_requests_on,
-    allow_playback_repeats = allow_playback_repeats,
     serialize_with = serialize_with,
     preserve_exact_body_bytes = preserve_exact_body_bytes,
     re_record_interval = re_record_interval,
-    clean_outdated_http_interactions = clean_outdated_http_interactions,
     warn_on_empty = warn_on_empty
   )
   cassette_push(cassette)
@@ -87,6 +83,9 @@ eject_cassette <- function() {
 #'   session.
 #' - `current_cassette()`: returns `NULL` when no cassettes are in use;
 #' returns the current cassette (a `Cassette` object) when one is in use
+#' - `currrent_cassette_recording()` and `current_cassette_replaying()`:
+#'   tell you if the current cassette is recording and/or replaying. They
+#'   both return `FALSE` if there is no cassette in use.
 #' - `cassette_path()`: returns the current directory path where cassettes
 #' will be stored
 #'
@@ -131,6 +130,26 @@ current_cassette <- function() {
     the$cassettes[[n]]
   } else {
     NULL
+  }
+}
+
+#' @export
+#' @rdname cassettes
+current_cassette_recording <- function() {
+  if (cassette_active()) {
+    current_cassette()$recording()
+  } else {
+    FALSE
+  }
+}
+
+#' @export
+#' @rdname cassettes
+current_cassette_replaying <- function() {
+  if (cassette_active()) {
+    current_cassette()$replaying()
+  } else {
+    FALSE
   }
 }