+++ R/request_handler.R
@@ -21,27 +21,21 @@ RequestHandler <- R6::R6Class(
         private$request_summary(self$request)
       )
 
-      if (private$externally_stubbed()) {
-        # FIXME: not quite sure what externally_stubbed is meant for
-        #   perhaps we can get rid of it here if only applicable in Ruby
-        vcr_log_sprintf("  externally stubbed")
-        return(private$on_externally_stubbed_request())
-      }
-
       if (should_be_ignored(self$request)) {
         vcr_log_sprintf("  ignored")
         return(private$on_ignored_request())
       }
 
-      if (casette_is_replayable()) {
-        interactions <- current_cassette()$http_interactions
+      if (current_cassette_replaying()) {
+        cassette <- current_cassette()
+        interactions <- cassette$http_interactions
         vcr_log_sprintf(
           "  Looking for existing requests using %s",
           paste0(interactions$request_matchers, collapse = "/")
         )
         idx <- interactions$find_request(self$request)
         if (!is.na(idx)) {
-          vcr_response <- interactions$response_for(idx)
+          vcr_response <- interactions$replay_request(idx)
           vcr_log_sprintf("  Replaying response %i", idx)
           return(private$on_stubbed_by_vcr_request(vcr_response))
         } else {
@@ -49,17 +43,25 @@ RequestHandler <- R6::R6Class(
         }
       }
 
-      if (cassette_is_recording()) {
+      if (current_cassette_recording()) {
         return(private$on_recordable_request())
       }
 
-      if (the$config$log) {
-        # Log messages already give the details
-        cli::cli_abort("Failed to find matching request in active cassette.")
-      } else {
-        err <- UnhandledHTTPRequestError$new(self$request)
-        err$run()
+      # Since it's going to error, there's no point in also giving a warning
+      # about the cassette being empty
+      if (cassette_active()) {
+        cassette <- current_cassette()
+        cassette$warn_on_empty <- FALSE
       }
+      cli::cli_abort(
+        c(
+          "Failed to find matching request in active cassette.",
+          i = if (!the$config$log)
+            "Use {.fn local_vcr_configure_log} to get more details.",
+          i = "Learn more in {.vignette vcr::debugging}."
+        ),
+        class = "vcr_unhandled"
+      )
     }
   ),
 
@@ -70,14 +72,12 @@ RequestHandler <- R6::R6Class(
     },
 
     # request type helpers
-    externally_stubbed = function() FALSE,
-
     get_stubbed_response = function(request) {
       if (!cassette_active()) {
         return(NULL)
       }
       interactions <- current_cassette()$http_interactions
-      interactions$response_for(request)
+      interactions$replay_request(request)
     },
 
     #####################################################################
@@ -86,8 +86,6 @@ RequestHandler <- R6::R6Class(
     ###   - all fxns take `request` param for consistentcy, even if they dont use it
     ##### so we can "monkey patch" these in each HTTP client adapter by
     #####   reassigning some of these functions with ones specific to the HTTP client
-
-    on_externally_stubbed_request = function() NULL,
 
     on_ignored_request = function() {
       # perform and return REAL http response
@@ -104,27 +102,3 @@ RequestHandler <- R6::R6Class(
     }
   )
 )
-
-cassette_is_recording <- function() {
-  if (cassette_active()) {
-    current_cassette()$recording()
-  } else {
-    FALSE
-  }
-}
-
-casette_is_replayable <- function() {
-  if (cassette_active()) {
-    current_cassette()$http_interactions$n_replayable() > 0
-  } else {
-    FALSE
-  }
-}
-
-cassette_has_response <- function(request) {
-  if (cassette_active()) {
-    current_cassette()$http_interactions$has_interaction(request)
-  } else {
-    FALSE
-  }
-}