+++ tests/testthat/test-ause_cassette_re_record.R
@@ -1,41 +1,32 @@
-conn <- crul::HttpClient$new(hb())
-
 test_that("use_cassette options: re_record_interval", {
-  local_vcr_configure(dir = withr::local_tempdir())
-
-  # first use
-  use_cassette(
-    "re_record1",
-    res <- conn$get("get"),
-    re_record_interval = 3L,
-    clean_outdated_http_interactions = TRUE
+  local_vcr_configure(
+    dir = withr::local_tempdir(),
+    re_record_interval = 1L
   )
-  rr1 <- read_cassette("re_record1.yml")
+
+  conn <- crul::HttpClient$new(hb())
+  # first use
+  use_cassette("test", res <- conn$get("get"))
+  rr1 <- read_cassette("test.yml")
 
   # second use, not expired, no change in recorded_at value
-  use_cassette(
-    "re_record1",
-    res <- conn$get("get"),
-    re_record_interval = 3L,
-    clean_outdated_http_interactions = TRUE
-  )
-  rr2 <- read_cassette("re_record1.yml")
+  use_cassette("test", res <- conn$get("get"))
+  rr2 <- read_cassette("test.yml")
 
   expect_equal(rr1$recorded_at, rr2$recorded_at)
 
   # third use, Sys.sleep, now expired, A change in recorded_at value
-  Sys.sleep(3L)
-  use_cassette(
-    "re_record1",
-    res <- conn$get("get"),
-    re_record_interval = 3L,
-    clean_outdated_http_interactions = TRUE
+  Sys.sleep(1.1)
+  local_vcr_configure_log(file = stdout())
+  expect_snapshot(
+    use_cassette("test", res <- conn$get("get")),
+    transform = function(x) {
+      x |>
+        gsub(hb(), "{httpbin}", x = _, fixed = TRUE) |>
+        gsub("\\d+ bytes", "{bytes} bytes", x = _)
+    }
   )
-  rr3 <- read_cassette("re_record1.yml")
-
-  # tests
-  # expect_equal(rr1$http_interactions[[1]]$re_record_interval, 10)
-  # expect_true(rr1$http_interactions[[1]]$clean_outdated_http_interactions)
+  rr3 <- read_cassette("test.yml")
 
   ## 1st and 2nd should be identical
   expect_true(