+++ tests/testthat/test-request_handler-httr2.R
@@ -52,6 +52,63 @@ test_that("can capture & replay raw body", {
   )
 })
 
+test_that("can send an image raw body in request - w/ appropriate header", {
+  local_vcr_configure(dir = withr::local_tempdir())
+  request <- httr2::request(hb("/post"))
+  img <- system.file(package = 'httr2', 'help/figures/logo.png')
+  bin <- readBin(img, what = 'raw', n = file.size(img))
+  request <- httr2::req_headers(request, 'Content-Type' = 'image/png')
+  request <- httr2::req_body_raw(request, bin)
+
+  # Record
+  use_cassette(
+    "test",
+    resp_record <- httr2::req_perform(request),
+    preserve_exact_body_bytes = TRUE
+  )
+  expect_s3_class(resp_record, "httr2_response")
+
+  # Replay
+  use_cassette(
+    "test",
+    resp_replay <- httr2::req_perform(request),
+    preserve_exact_body_bytes = TRUE
+  )
+  expect_s3_class(resp_replay, "httr2_response")
+  expect_equal(
+    httr2::resp_body_raw(resp_replay),
+    httr2::resp_body_raw(resp_record)
+  )
+})
+
+test_that("can send an image raw body in request - w/o appropriate header", {
+  local_vcr_configure(dir = withr::local_tempdir())
+  request <- httr2::request(hb("/post"))
+  img <- system.file(package = 'httr2', 'help/figures/logo.png')
+  bin <- readBin(img, what = 'raw', n = file.size(img))
+  request <- httr2::req_body_raw(request, bin)
+
+  # Record
+  use_cassette(
+    "test",
+    resp_record <- httr2::req_perform(request),
+    preserve_exact_body_bytes = TRUE
+  )
+  expect_s3_class(resp_record, "httr2_response")
+
+  # Replay
+  use_cassette(
+    "test",
+    resp_replay <- httr2::req_perform(request),
+    preserve_exact_body_bytes = TRUE
+  )
+  expect_s3_class(resp_replay, "httr2_response")
+  expect_equal(
+    httr2::resp_body_raw(resp_replay),
+    httr2::resp_body_raw(resp_record)
+  )
+})
+
 test_that("captured suppressed error", {
   local_vcr_configure(dir = withr::local_tempdir())
   request <- httr2::request(hb("/status/404"))
@@ -96,12 +153,7 @@ test_that("httr2 redacts auth header", {
   request <- httr2::request(hb("/basic-auth/foo/bar"))
   request <- httr2::req_auth_basic(request, "foo", "bar")
   use_cassette("test", response <- httr2::req_perform(request))
-
-  yml <- read_cassette("test.yml")
-  expect_equal(
-    yml$http_interactions[[1]]$request$headers,
-    list(Authorization = "<redacted>")
-  )
+  expect_equal(vcr_last_request()$headers, list(Authorization = "<redacted>"))
 })
 
 test_that("can capture body: string", {
@@ -118,9 +170,7 @@ test_that("can capture body: string", {
     httr2::resp_body_json(resp_record),
     httr2::resp_body_json(resp_replay)
   )
-
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_equal(interaction$request$body$string, "body")
+  expect_equal(vcr_last_request()$body, list(string = "body"))
 })
 
 test_that("binary body uses base64 encoding", {
@@ -131,8 +181,7 @@ test_that("binary body uses base64 encoding", {
   req <- httr2::req_headers(req, "Accept" = "image/png")
 
   use_cassette("test", httr2::req_perform(req))
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_named(interaction$response$body, "raw_gzip")
+  expect_named(vcr_last_response()$body, "raw_gzip")
 })
 
 test_that("can capture body: json", {
@@ -149,9 +198,7 @@ test_that("can capture body: json", {
     httr2::resp_body_json(resp_record),
     httr2::resp_body_json(resp_replay)
   )
-
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_equal(interaction$request$body$string, "{\"foo\":\"bar\"}")
+  expect_equal(vcr_last_request()$body, list(string = "{\"foo\":\"bar\"}"))
 })
 
 test_that("can capture body: form", {
@@ -168,9 +215,7 @@ test_that("can capture body: form", {
     httr2::resp_body_json(resp_record),
     httr2::resp_body_json(resp_replay)
   )
-
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_equal(interaction$request$body$string, "a=x&b=y")
+  expect_equal(vcr_last_request()$body, list(string = "a=x&b=y"))
 })
 
 test_that("can capture body: multipart", {
@@ -187,9 +232,7 @@ test_that("can capture body: multipart", {
     httr2::resp_body_json(resp_record),
     httr2::resp_body_json(resp_replay)
   )
-
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_equal(interaction$request$body$fields, list(a = "x", b = "y"))
+  expect_equal(vcr_last_request()$body, list(fields = list(a = "x", b = "y")))
 })
 
 test_that("can capture body: file", {
@@ -207,9 +250,7 @@ test_that("can capture body: file", {
     httr2::resp_body_json(resp_record),
     httr2::resp_body_json(resp_replay)
   )
-
-  interaction <- read_cassette("test.yml")$http_interactions[[1]]
-  expect_equal(interaction$request$body$string, path)
+  expect_equal(vcr_last_request()$body, list(string = path))
 })
 
 test_that("redacted headers handled appropriately", {
@@ -225,8 +266,8 @@ test_that("redacted headers handled appropriately", {
       httr2::req_perform()
   })
 
-  cas <- read_cassette("redacted_httr2.yml")
-  headers <- cas$http_interactions[[1]]$request$headers
-  expect_equal(headers$NotASecret, "NotHidden")
-  expect_equal(headers$SecretHeader, "<redacted>")
+  expect_equal(
+    vcr_last_request()$headers,
+    list(NotASecret = "NotHidden", SecretHeader = "<redacted>")
+  )
 })